move random_data to custom_ingest
