+++ hooli_data_eng/custom_ingest/resources.py
@@ -10,17 +10,83 @@
 import responses
 import requests
 import pandas as pd
-from hooli_data_eng.utils import random_data
 import numpy as np
 import random
+import datetime
+import time
+import uuid
 import dagster as dg
-from typing import Optional
-
+from typing import Optional, Any, Dict
 from hooli_data_eng.resources import (
     snowflake_branch_io_manager,
     snowflake_prod_io_manager,
     duckdb_io_manager,
 )
+
+
+def _random_times(n: int):
+    """Generate some random times that generally become more frequent as time goes on."""
+    time.sleep(0.5)
+    start = pd.to_datetime("2022-01-01")
+    end = pd.to_datetime(datetime.datetime.now())
+
+    start_u = start.value // 10**9
+    end_u = end.value // 10**9
+
+    dist = np.random.standard_exponential(size=n) / 10
+
+    clipped_flipped_dist = 1 - dist[dist <= 1]
+    clipped_flipped_dist = clipped_flipped_dist[:-1]
+
+    if len(clipped_flipped_dist) < n:
+        clipped_flipped_dist = np.append(
+            clipped_flipped_dist, clipped_flipped_dist[: n - len(clipped_flipped_dist)]
+        )
+
+    times = pd.to_datetime(
+        (clipped_flipped_dist * (end_u - start_u)) + start_u, unit="s"
+    )
+
+    hours = times.round("60min").to_pydatetime()
+
+    return hours
+
+
+def random_data(
+    extra_columns: Dict[str, Any], n: int, filter_date=None
+) -> pd.DataFrame:
+    skus = [
+        "pepsi",
+        "coke",
+        "sprite",
+        "coke zero",
+        "powerade",
+        "diet",
+        "gingerale",
+        "juice",
+    ]
+
+    # always have user_id
+    data = {"user_id": np.random.randint(0, 1000, size=n)}
+
+    for name, dtype in extra_columns.items():
+        if name == "sku":
+            data[name] = random.choices(skus, k=n)
+        elif dtype is str:
+            data[name] = [str(uuid.uuid4()) for _ in range(n)]
+        elif dtype is int:
+            data[name] = np.random.randint(0, 100, size=n)
+        elif dtype is float:
+            data[name] = 100 * np.random.random(size=n)
+
+    data = pd.DataFrame(data)
+
+    if filter_date:
+        data["dt"] = pd.to_datetime(filter_date)
+    else:
+        data["dt"] = _random_times(n=n)
+
+    return pd.DataFrame(data)
 
 
 class RawDataAPI(dg.ConfigurableResource):
